add GitHub action to rebuild the API
The create-pull-request action only fires off a PR when there's actually changes to the
repo. And rebuild.sh will only emit changes when a new Go release is present or
when we change the files outputted by it.

We run this action every 3 hours or when we push to the force_rebuild_api
branch.